requirements.txt
# general modules
django==3.2
python-decouple
pillow
djangorestframework
# third party modules
markdown
django-filter
coreapi
drf-yasg[validation]
rest_framework_simplejwt
djoser
# deployment moduls

# email third party modules
django-mail-templated

black
flake8